% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1B1.R
\name{fars_map_state}
\alias{fars_map_state}
\title{name
name fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{numeric}

\item{year}{numeric}
}
\description{
TITLE
title function to draw a map
}
\details{
#’  Function to check if there is data for a state in a particular year file. If data
is present and plot longitude against latitude for observations where
longitude is > 900 and latitude is >90
Add Points to a Plot
}
\note{
-fars_map_state- draw a map for a given year and state
}
\examples{
\dontrun{fars_map_state(20,2015)}
}
