% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1B1.R
\name{fars_read}
\alias{fars_read}
\title{NAME
name fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{-}{filename : a character string with path of the file}
}
\value{
- Tibble object with csv file data
}
\description{
TITLE
title read function
#'@note
fars_functions : Package for reading, summarising and create visual representation
for data from the Fatality Analysis Reporting System (FARS).
}
\details{
: Function Checks the existence of given file. In case file is not found, a
          message is generated. While reading the file,suppresses the progress
          messages to have a clean output
}
\note{
fars_read function reads the given file and convert resulting data frame
to a table using tbl_df function from dplyr
}
\examples{
\dontrun{- fars_read("accident_2015.csv.bz2")}
}
