% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1B1.R
\name{fars_read_years}
\alias{fars_read_years}
\title{name
name fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{-}{a numeric vector contains years e.g. c(2015,2001,1980)}
}
\value{
- a list of tibbles each containing year and month columns. If files does not exist, returns NULL
}
\description{
TITLE
title function to iteratively read the files
}
\note{
-fars_read_years function to iteratively read the files based on numeric year
vector
}
\examples{
\dontrun{fars_read_years(c(2015,1980,2010))}
}
