% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1B1.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{name
name fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{numeric}
}
\value{
A wide format tibble. One column for year and a MONTH column. Each
value is the count of observations in a year-month.
}
\description{
TITLE
title combines files
}
\note{
-fars_summarize_years combines multiple year files by rows and then group on year and month
Further it calculates summary of the resulting data frame and spreads across year
}
\examples{
\dontrun{fars_summarize_years(c(2013,2014,2015))}
}
