% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1B1.R
\name{make_filename}
\alias{make_filename}
\title{NAME
name make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{-}{year- Numeric e.g. 2015}
}
\value{
- character string having custom file name e.g.accident_2015.csv.bz2
}
\description{
TITLE
title function creates a custom file name based on year argument value
}
\note{
make_filename function creates a custom file name based on year argument value
}
\examples{
\dontrun{make_filename(2015)}
}
